+++ core/fileformats/objLoader.cpp
@@ -82,17 +82,17 @@ int OBJLoader::loadOBJ(istream &input, Mesh3DDecorated &mesh)
                 for (int j = 0; j < 3 && std::getline(splitter, part, '/'); j++)
                 {
                     if (j == 0) {
-                        size_t id = std::stoi(part);
+                        int id = std::stoi(part);
                         face[i] = id - 1;
                     }
 
                     if (j == 1) {
-                        size_t id = std::stoi(part);
+                        int id = std::stoi(part);
                         texId[i] = id - 1;
                     }
 
                     if (j == 2) {
-                        size_t id = std::stoi(part);
+                        int id = std::stoi(part);
                         normId[i] = id - 1;
                     }
                 }
@@ -164,7 +164,7 @@ int OBJLoader::loadOBJSimple(istream &input, Mesh3D &mesh)
 
                 for (int j = 0; j < 1 && std::getline(splitter, part, '/'); j++)
                 {
-                    size_t id = std::stoi(part);
+                    int id = std::stoi(part);
                     face[i] = id - 1;
                 }
             }
